# refactor `hashbuilder::update` 