just some minor changes