@malik672 2025-04-02
> can you please elaborate a bit on the changes, otherwise I'm having a hard time reviewing this

Ohh, optimized the HashBuilder update process by reducing redundant operations, improving memory management, and refining key assertions. These changes result in a ~21.8% reduction in update latency (32.81ms → 25.7ms), as observed on a  flame graph.

 Improvements
 - Optimized Default Call

`self.state_masks.resize(len, TrieMask::default())`
`self.state_masks.resize(len, *trie_default)`
`while self.state_masks.last() == Some(trie_default) {
    self.state_masks.pop();
}
`
basically you store the default value once before entering the loop and using it multiple times in the loop instead of making a default call each time in the loop

- Refactored add_leaf for Efficiency(not really sure about this)
basically I did this to avoid additional call frame in a loop(should benchmark this in reth) and the _key part is just storing then reusing
let _key = &self.key;
`assert!(&key > _key, "add_leaf key {:?} self.key {:?}", key, _key)`
this one is related to this: https://github.com/paradigmxyz/reth/issues/15398?notification_referrer_id=NT_kwDOA6iyUrQxNTU5NDEyOTU3Njo2MTM4NzM0Ng#issuecomment-2765277755


- Improved Branching Logic
`let s_and_p = succeeding.is_empty() || preceding_exists`
Consolidates multiple conditional checks, reducing branching overhead

- Memory & Safety Enhancements
Ensured `current_len > 0` before unsafe indexing:
`TrieMask::from_nibble(*unsafe { current.get_unchecked(sub_len) })`

- storing `current.slice(..len_from)` then reusing it again instead of calling it twice in quick succession
technically this brought the highest change

Performance Gains
Update Latency: 32.81ms → 25.7ms (~21.8% improvement).