@malik672 2025-04-02
in all these are just micro optimizations, I think any significant one will be trying to eliminate `resize` and that's hardly viable