@malik672 2025-04-02
> > Ohh, optimized the HashBuilder update process by reducing redundant operations, improving memory management, and refining key assertions. These changes result in a ~21.8% reduction in update latency (32.81ms → 25.7ms), as observed on a flame graph.
> 
> Which benchmark is this?

flamegraph

criterion is 725ms to 712ms